Add a simple HTTP interface to the logger plugin.
See http://localhost:8080/logger. This is super-basic: just a list of
servers, of channels logged on each server, of the log files for a
particular channel, and an interface to return those files when
requested.

Tries to be robust about evil input; treats the bot config as the source
of truth as much as possible, and only allows access to files which have
a configured log directory. Only touches the filesystem in configured
log directories, and never uses arbitrary filenames.